# src/main/resources/static/registry.openapi.yaml
openapi: 3.0.0
info:
  title: docker registry
  version: 1.0
servers:
  - url: http://192.168.50.28:5000/v2
  - url: https://index.docker.io/v2
paths:
...
  /{name}/manifests/{reference}:
    get:
      description: Pulling an Image Manifest
      operationId: getManifests
      parameters:
        - in: path
          name: name
          schema:
            type: string
        - in: path
          name: reference
          schema:
            type: string
        - in: header
          name: Authorization
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2ManifestResult'
    delete:
      description: delete image
      operationId: deleteManifests
      parameters:
        - in: path
          name: name
          schema:
            type: string
        - in: path
          name: reference
          schema:
            type: string
      responses:
        '202':
          description: Accept
components:
  schemas:
...
    V2ManifestResult:
      type: object
      properties:
        schemaVersion:
          type: integer
        mediaType:
          type: string
        config:
          $ref: '#/components/schemas/Manifest'
        layers:
          type: array
          items:
            $ref: '#/components/schemas/Manifest'
    Manifest:
      type: object
      properties:
        mediaType:
          type: string
        size:
          type: integer
          format: int64
        digest:
          type: string
...